Add live subagent progress pane and clean lint

+++ src/core/skills.h
@@ -13,7 +13,8 @@ namespace pi::core {
 inline constexpr size_t kSkillMaxNameLength = 64;
 inline constexpr size_t kSkillMaxDescriptionLength = 1024;
 inline constexpr size_t kSkillMaxShortDescriptionLength = 256;
-inline constexpr size_t kSkillMaxBodyBytes = 128 * 1024;
+inline constexpr size_t kSkillMaxBodyBytes =
+    static_cast<const size_t>(128 * 1024);
 inline constexpr size_t kSkillMaxPerRoot = 256;
 inline constexpr size_t kSkillMaxTotal = 512;
